+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 1.1.0
 
 - Added Security util for url signing.
+- Added support for custom domain and worker url